refactor: 두번째 미션 마지막 피드백 적용
