+++ src/main/java/kitchenpos/menus/tobe/domain/common/CleanName.java
@@ -1,7 +1,8 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.common;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
+import kitchenpos.menus.tobe.domain.MenuNameValidationService;
 
 import java.util.Objects;
