+++ src/test/java/kitchenpos/menus/domain/CleanNameTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.domain;
 
 import kitchenpos.menus.application.FakeMenuPurgomalumClient;
-import kitchenpos.menus.tobe.domain.CleanName;
+import kitchenpos.menus.tobe.domain.common.CleanName;
 import kitchenpos.menus.tobe.domain.MenuNameValidationService;
 import kitchenpos.products.infra.PurgomalumClient;
 import org.assertj.core.api.Assertions;
