+++ src/test/java/kitchenpos/menus/domain/PriceTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.menus.domain;
 
-import kitchenpos.menus.tobe.domain.Price;
+import kitchenpos.menus.tobe.domain.common.Price;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,7 +13,7 @@ public class PriceTest {
     @Test
     @DisplayName("동일한 가격인지 비교한다")
     void equals() {
-        Price 메뉴_상품_가격 = new Price(BigDecimal.valueOf(20_000), 5);
+        Price 메뉴_상품_가격 = new Price(BigDecimal.valueOf(20_000)).multiplyBy(5);
         Price 십만원 = new Price(BigDecimal.valueOf(100_000));
 
         Assertions.assertThat(메뉴_상품_가격.equals(십만원)).isTrue();
